+++ R/evaluator-deprecated.R
@@ -1,0 +1,8 @@
+#' @title Deprecated functions in package \pkg{evaluator}.
+#' @description The functions listed below are deprecated and will be defunct in
+#'   the near future. When possible, alternative functions with similar
+#'   functionality are also mentioned. Help pages for deprecated functions are
+#'   available at \code{help("<function>-deprecated")}.
+#' @name evaluator-deprecated
+#' @keywords internal
+NULL