+++ R/evaluator-package.R
@@ -6,3 +6,7 @@
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
+
+## quiets concerns of R CMD check re: the .'s that appear in pipelines
+## technique from Jenny Bryan's googlesheet package
+utils::globalVariables(c("."))